+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 - Clustering no longer panics when no genomes pass the quality thresholds, instead warning and generating empty output ([#75](https://github.com/wwood/galah/issues/75), reported by [@erikrikarddaniel](https://github.com/erikrikarddaniel))
+- fastANI is no longer required to be installed when clustering with skani, and skani's presence is now verified when it is used
 
 ## Version 0.5.2